Dissolved the redux folder
