+++ src/client/src/appBootstrapper.tsx
@@ -1,11 +1,11 @@
 import 'babel-polyfill'
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import { User } from './types'
-import createConnection from './system/service/connection'
-import { OpenFin } from './system/openFin'
-import { ShellContainer, OpenFinProvider } from './ui/shell'
 import { Provider } from 'react-redux'
+import { OpenFin } from './system/openFin'
+import createConnection from './system/service/connection'
+import { User } from './types'
+import { ShellContainer, OpenFinProvider } from './ui/shell'
 // TODO: change to import when webpack bug solved https://github.com/webpack/webpack/issues/4160
 const config = require('config.json')
 
@@ -26,12 +26,8 @@ declare const window: any
 const connectSocket = () => {
   const user: User = FakeUserRepository.currentUser
   const realm = 'com.weareadaptive.reactivetrader'
-  const url = config.overwriteServerEndpoint
-    ? config.serverEndPointUrl
-    : location.hostname
-  const port = config.overwriteServerEndpoint
-    ? config.serverPort
-    : location.port
+  const url = config.overwriteServerEndpoint ? config.serverEndPointUrl : location.hostname
+  const port = config.overwriteServerEndpoint ? config.serverPort : location.port
   return createConnection(user.code, url, realm, port)
 }
