+++ src/client/src/combineReducers.ts
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux'
+import compositeStatusServiceReducer from './compositeStatusServiceOperations'
 import connectionStatusReducer from './connectionStatusOperations'
-import compositeStatusServiceReducer from './redux/reducers/compositeStatusServiceOperations'
-import { currencyPairReducer } from './redux/reducers/currencyPairsReducer'
-import { pricingServiceReducer } from './redux/reducers/pricingReducer'
+import { currencyPairReducer } from './currencyPairsOperations'
+import { pricingServiceReducer } from './pricingOperations'
 import { analyticsReducer } from './ui/analytics/index'
 import { blotterReducer } from './ui/blotter/index'
 import { regionsReducer } from './ui/common/regions/regionsOperations'
