+++ src/client/src/configureStore.ts
@@ -4,10 +4,10 @@ import { createEpicMiddleware, combineEpics } from 'redux-observable'
 import { persistStore } from 'redux-persist'
 
 import rootReducer from './combineReducers'
+import { compositeStatusServiceEpic } from './compositeStatusServiceOperations'
 import { connectionStatusEpicsCreator } from './connectionStatusOperations'
-import { pricingServiceEpic } from './redux/actions/pricingActions'
-import { referenceServiceEpic } from './redux/actions/referenceDataActions'
-import { compositeStatusServiceEpic } from './redux/reducers/compositeStatusServiceOperations'
+import { pricingServiceEpic } from './pricingOperations'
+import { referenceServiceEpic } from './referenceDataOperations'
 import { analyticsServiceEpic } from './ui/analytics'
 import { blotterEpic } from './ui/blotter/'
 import { popoutEpic } from './ui/common/popout/popoutEpic'
