+++ src/client/src/currencyPairsOperations.ts
@@ -1,5 +1,5 @@
-import { CurrencyPair } from '../../types/index'
-import { ACTION_TYPES } from '../actions/referenceDataActions'
+import { CurrencyPair } from './types/index'
+import { ACTION_TYPES } from './referenceDataOperations'
 
 interface CurrencyPairReducerState {
   [id: string]: CurrencyPair
@@ -10,7 +10,7 @@ export const currencyPairReducer = (state: CurrencyPairReducerState = {}, action
     case ACTION_TYPES.REFERENCE_SERVICE:
       const payloadUpdateItems = action.payload.currencyPairUpdates
       const updatedItems = {}
-      payloadUpdateItems.forEach((item) => {
+      payloadUpdateItems.forEach(item => {
         updatedItems[item.currencyPair.symbol] = item.currencyPair
       })
       return { ...state, ...updatedItems }
